refactor(cli): extract retrain commands to cli/_retrain.py (#284, step 3) (#393)
Move `vstash retrain` and `vstash retrain-multi` (~670 lines) into
vstash/cli/_retrain.py; __init__.py imports the module for its `@app.command()`
side effects. Establishes the command-group module pattern: each module imports
`app` + the shared helpers from `._app`, defines its commands, and is imported
by __init__ to register them.

Bodies moved verbatim. __init__.py drops the imports retrain was the last user
of (click, VstashStore, _load_labeled_queries_jsonl). The retrain test fixture
patches `_get_store` on cli._retrain (where the command reads it) instead of cli.

- pytest tests/ -> 1290 passed; ruff clean; retrain/-multi resolve via CliRunner
